+++ lib/logstash/outputs/elasticsearch/http_client_builder.rb
@@ -4,6 +4,7 @@ module LogStash; module Outputs; class ElasticSearch;
       client_settings = {
         :pool_max => params["pool_max"],
         :pool_max_per_route => params["pool_max_per_route"],
+        :check_connection_timeout => params["validate_after_inactivity"]
       }
 
       common_options = {
@@ -21,6 +22,10 @@ module LogStash; module Outputs; class ElasticSearch;
 
       if params["path"]
         client_settings[:path] = "/#{params["path"]}/".gsub(/\/+/, "/") # Normalize slashes
+      end
+
+      if params["parameters"]
+        client_settings[:parameters] = params["parameters"]
       end
 
       logger.debug? && logger.debug("Normalizing http path", :path => params["path"], :normalized => client_settings[:path])