+++ travis-run.sh
@@ -10,9 +10,12 @@ trap finish EXIT
 
 setup_es() {
   download_url=$1
-  curl -sL $download_url > elasticsearch.tar.gz
-  mkdir elasticsearch
-  tar -xzf elasticsearch.tar.gz --strip-components=1 -C ./elasticsearch/.
+  if [[ ! -d elasticsearch ]]; then
+    curl -sL $download_url > elasticsearch.tar.gz
+    mkdir elasticsearch
+    tar -xzf elasticsearch.tar.gz --strip-components=1 -C ./elasticsearch/.
+  fi
+  rm -f elasticsearch/config/scripts || true
   ln -sn ../../spec/fixtures/scripts elasticsearch/config/.
 }
 
@@ -33,17 +36,25 @@ start_es() {
 if [[ "$INTEGRATION" != "true" ]]; then
   bundle exec rspec -fd spec
 else
+  if [ "$1" -eq "" ]; then
+    spec_path="spec"
+  else
+    spec_path="$1"
+  fi
   if [[ "$ES_VERSION" == 5.* ]]; then
-    setup_es https://download.elastic.co/elasticsearch/release/org/elasticsearch/distribution/tar/elasticsearch/$ES_VERSION/elasticsearch-$ES_VERSION.tar.gz
-    start_es -Ees.script.inline=true -Ees.script.indexed=true -Ees.script.file=true
-    bundle exec rspec -fd spec --tag integration --tag version_5x --tag integration_2x_plus
+    setup_es https://artifacts.elastic.co/downloads/elasticsearch/elasticsearch-${ES_VERSION}.tar.gz
+    start_es -Escript.inline=true -Escript.stored=true -Escript.file=true
+    # Run all tests which are for versions > 5 but don't run ones tagged < 5.x. Include ingest, new template
+    bundle exec rspec -fd --tag integration --tag version_greater_than_equal_to_5x --tag ~version_less_than_5x $spec_path
   elif [[ "$ES_VERSION" == 2.* ]]; then
     setup_es https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-$ES_VERSION.tar.gz
     start_es -Des.script.inline=on -Des.script.indexed=on -Des.script.file=on
-    bundle exec rspec -fd spec --tag integration --tag ~version_5x
+    # Run all tests which are for versions < 5 but don't run ones tagged 5.x and above. Skip ingest, new template
+    bundle exec rspec -fd --tag integration --tag version_less_than_5x --tag ~version_greater_than_equal_to_5x $spec_path
   else
     setup_es https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-$ES_VERSION.tar.gz
     start_es -Des.script.inline=on -Des.script.indexed=on -Des.script.file=on
-    bundle exec rspec -fd spec --tag integration --tag ~version_5x --tag ~version_2x_plus
+    # Still have to support ES versions < 2.x so run tests for those.
+    bundle exec rspec -fd --tag integration --tag ~version_greater_than_equal_to_5x --tag ~version_greater_than_equal_to_2x $spec_path
   fi
 fi